알림 기능 구현 (#7)
* chore: build.gradle

* error: comment

* feat: 게시글 좋아요 알림 기능#1

* design: 모달창 활용하여 알림 리스트 UI 구현

* feat: 알림 기능#2, eventListener

* chore: common.css 및 자잘한 코드 수정

* fix: 모달창 스크롤 적용 안되는 문제 해결

* fix: 좋아요 개수 바로 안보이는 문제 해결

* refactor: ajax 활용해서 조회수 카운트

* chore: 텍스트 넘침 문제 해결

* chore: 텍스트 넘침 문제 해결#2

+++ src/main/java/com/kl/grooveo/boundedContext/reply/controller/FreedomPostReplyController.java
@@ -14,10 +14,13 @@ import com.kl.grooveo.boundedContext.reply.service.FreedomPostReplyService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 @RequestMapping("/reply")
 @RequiredArgsConstructor
@@ -29,6 +32,7 @@ public class FreedomPostReplyController {
     private final MemberService memberService;
     private final Rq rq;
 
+    @PreAuthorize("isAuthenticated()")
     @PostMapping("/create/{id}")
     public String create(Model model, @PathVariable("id") Long id,
                          @Valid ReplyForm replyForm, @RequestParam(value = "commentPage", defaultValue = "0") int commentPage,
@@ -55,15 +59,16 @@ public class FreedomPostReplyController {
                 freedomPostReply.getFreedomPostComment().getFreedomPost().getId(), commentPage, freedomPostReply.getFreedomPostComment().getId());
     }
 
+    @PreAuthorize("isAuthenticated()")
     @DeleteMapping("/{id}")
     public String delete(@PathVariable("id") Long id) {
         FreedomPostReply freedomPostReply = this.freedomPostReplyService.getReply(id);
         FreedomPostComment freedomPostComment = this.freedomPostCommentService.getComment(freedomPostReply.getFreedomPostComment().getId());
 
-//        if (!freedomPostReply.getAuthor().getUsername().equals(principal.getName())
-//                && !freedomPostComment.getAuthor().getUsername().equals(principal.getName())) {
-//            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "삭제권한이 없습니다.");
-//        }
+        if (!freedomPostReply.getAuthor().getUsername().equals(rq.getMember().getUsername())
+                && !freedomPostComment.getAuthor().getUsername().equals(rq.getMember().getUsername())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "삭제권한이 없습니다.");
+        }
 
         this.freedomPostReplyService.delete(freedomPostReply);
         return String.format("redirect:/community/freedomPost/detail/%s", freedomPostComment.getFreedomPost().getId());
